Turn the NDK tools into a real python package.
Using the test runner on Windows requires a lot of manual setup for
pulling the various dependencies. If we build a real package we can
just ship that instead, and this also lets us clean up all our uses
of site.addsitedirectory.

The new ./run_tests.py is a shortcut as would be installed by pip. We
need this to continue supported the non-installed workflow since the
build servers still follow that.

Test: ./checkbuild.py && ./run_tests.py
Test:
    ./checkbuild.py --no-build-tests
    mkdir foo
    cd foo
    virtualenv .env
    . .env/activate
    pip install ../../out/dist/adb-0.0.1.tar.gz
    pip install ../../out/dist/lit-0.5.0.dev0.tar.gz
    pip install ../../out/dist/ndk-1.0.0.tar.gz
    run_tests.py \
        --rebuild \
        --config ../qa_config.json \
        --ndk ../../out/android-ndk-r17-canary \
        --test-src ../tests \
        out
Bug: None

Change-Id: I36080a0565a70467ab4ede15a66a3a34af224e90