ndk/win32.py: declare all Win32 functions the program uses
Previously, GetCurrentProcess returned -1 (a Python int),
and passing this -1 to AssignProcessToJobObject failed with an
error 6 (The handle is invalid).

If I then declare GetCurrentProcess as returning HANDLE, then
it returns UINT64_MAX (as a Python long). Passing it to
AssignProcessToJobObject then fails with this error:

ctypes.ArgumentError: argument 2: <type 'exceptions.OverflowError'>:
long int too long to convert

We need to be explicit about 64-bit return types and parameter
types. This patch declares all of them for consistency.

For whatever reason, the initial invalid handle error happens with
the Python 2.7.3 preinstalled on my Windows laptop but not with the
prebuilt NDK Python 2.7.

Test: ./run_tests.py --rebuild (on Windows)
Test: ./run_tests.py (on Linux)
Change-Id: I7e1917da77b5882f129d090e6d960e3a26fb4af4