Move APIs from build.lib.build_support to ndk.
"build" is not a valid package name for setuptools. This package will
be silently removed from the source distribution because setuptools
thinks it's the build directory rather than a python package named
build. Pieces of this package are being moved into the ndk package
where they belong, but will continue to be exported from there until
we can chase down all the other users.

Test: ./checkbuild.py && ./run_tests.py
Bug: None
Change-Id: I6c57eefb78866a6e4f53af22791842e1afb24698