parse_elfnote: rewrite, handle small sections
The .note.android.ident section includes a platform ABI version integer,
both in platform binaries and NDK-generated binaries. Starting with NDK
r14, the section also contains the NDK version and build number.

This script was printing garbage on binaries without the two NDK fields:

$ python parse_elfnote.py ~/android-sdk/lldb/2.3/android/armeabi/lldb-server
    ----------ABI INFO----------
    ABI_NOTETYPE: 1
    ABI_VENDOR: Android
    ABI_ANDROID_API: 21
    ABI_NDK_VERSION: ... garbage ...
    ABI_NDK_BUILD_NUMBER: ... garbage ...

This commit hides those two fields when they aren't present. It also hides
the first two fields, because they're redundant. (Either they're Android/1,
or there's an unexpected kind of note in our .note.android.ident section,
and we don't know how to decode it.)

New output:

$ ./parse_elfnote.py ~/android-sdk/lldb/2.3/android/armeabi/lldb-server
----------ABI INFO----------
ABI_ANDROID_API: 21

$ ./parse_elfnote.py ~/android-sdk/lldb/3.0/android/armeabi/lldb-server
----------ABI INFO----------
ABI_ANDROID_API: 21
ABI_NDK_VERSION: r14b
ABI_NDK_BUILD_NUMBER: 3816874

Details:

The .note.android.ident section is a NOTE section
(e.g. http://www.netbsd.org/docs/kernel/elf-notes.html) generated by
crtbrand.S. It needs to follow the NOTE format at least for the sake of
`readelf -n` (and AFAICT it does).

A generic ELF note has this format:
 - int32 namesz
 - int32 descsz
 - int32 notetype
 - char name[roundup(namesz, 4)]  (optional and variably-sized)
 - char desc[roundup(descsz, 4)]  (optional and variably-sized)

Previously, the script printed the NUL-stripped name, but it ignored
namesz, which specifies the length of the name field, and assumed the name
was 8 bytes long. It printed name/notetype along with the decoded desc, but
the point of name/notetype is to specify the format of the descriptor. If
it isn't Android/1, we shouldn't attempt to decode it.

This change parses a generic ELF note section, then parses the special
Android/1 note (in either the basic or NDK-extended format).

namesz counts a NUL-terminator, but it is also allowed to equal 0
(indicating a lack of a name rather than an empty name). (This script
ignores that distinction.)

A generic note section can have multiple notes, but our section has only
the one.

This change also makes the script executable.

Test: manual
Change-Id: I89190c9c14d5d0db7616c09138cf5a7a8633faa9