DummyWorkQueue: Add workers() for use by get_ui_lines
I added this while trying to debug a checkbuild.py hang. The hang was
apparently happening in the tcsetattr call, and has apparently stopped
reproducing now.

To debug: switch workqueue.WorkQueue to DummyWorkQueue. Without this
change, the build finishes immediately, and an exception is logged to the
first build file in out/dist/logs.

Test: manual
Change-Id: Ia22a2473d96a1ad5c1bedefbedd61ea91a78d0b1