MinGW: Set CC/CXX/... using only the basename.
The MinGW prebuilt only has the 64-bit commands, not the 32-bit ones. The
32-bit commands are created by prepare_canadian_toolchain, which is called
even for ordinary non-Canadian cross-compilation. We don't know the
build directory that will be passed to prepare_canadian_toolchain, so we
can't use an absolute path for the 32-bit build.

Reverts the second half of Ibf6fddba0f91739fabcc5c183f35ed436864b58a.

Bug: https://github.com/android-ndk/ndk/issues/624
Test: ./checkbuild.py --system=window
Test: ./checkbuild.py --system=windows64
Change-Id: I9de185210f1d2c4b0e16094b5a42aef8b3c70ac3