Fix bug in ldflags_to_sanitizers.py.
The Python tests were passing because we only tested the individual
functions, not the program as a whole.

The run_tests.py tests were passing because errors from $(shell) do
not cause the build to fail; ndk-build just ended up parsing the
error message and not finding any sanitizers.

The asan-smoke test passed because A) I hadn't run `./run_tests.py
--clean-device` to purge the old tests, so the libraries were still
on the device and B) running the tests in the r17 branch had led to
asan_device_setup running again, installing the libraries to the
system partition once again.

Test: nose2 build
Test: ./run_tests.py --clean-device  # After purging asan /system libs
Bug: https://github.com/android-ndk/ndk/issues/540
Change-Id: I85a30cc198be4a0f39548271f56112835805ca76